+++ scripts/hooks.js
@@ -13,16 +13,17 @@ Hooks.once('setup', () => {
     default: true,
   });
 
-  if (game.system.id === 'dnd5e') {
-    game.settings.register(Stealthy.MODULE_ID, 'ignoreFriendlyUmbralSight', {
-      name: game.i18n.localize("stealthy.dnd5e.ignoreFriendlyUmbralSight.name"),
-      hint: game.i18n.localize("stealthy.dnd5e.ignoreFriendlyUmbralSight.hint"),
-      scope: 'world',
-      config: true,
-      type: Boolean,
-      default: false,
-    });
-  }
+  game.settings.register(Stealthy.MODULE_ID, 'spotSecretDoors', {
+    name: game.i18n.localize("stealthy.spotSecretDoors.name"),
+    hint: game.i18n.localize("stealthy.spotSecretDoors.hint"),
+    scope: 'world',
+    config: true,
+    type: Boolean,
+    default: false,
+    onChange: value => {
+      debouncedReload();
+    },
+  });
 
   let sources = {
     'none': game.i18n.localize("stealthy.source.min"),