+++ scripts/stealthy.js
@@ -42,6 +42,110 @@ export class StealthyBaseEngine {
       libWrapper.MIXED,
       { perf_mode: libWrapper.PERF_FAST }
     );
+
+    if (game.settings.get(Stealthy.MODULE_ID, 'spotSecretDoors')) {
+      libWrapper.register(
+        Stealthy.MODULE_ID,
+        'Wall.prototype.createDoorControl',
+        function (wrapped) {
+          return StealthyBaseEngine.BaseWallCreateDoorControlSansGmCheck(this);
+        },
+        libWrapper.OVERRIDE,
+        { perf_mode: libWrapper.PERF_AUTO }
+      );
+
+      libWrapper.register(
+        Stealthy.MODULE_ID,
+        'DoorControl.prototype.isVisible',
+        function (wrapped) {
+          Stealthy.log('DoorControl.prototype.isVisible', this);
+
+          const wallDoc = this.wall.document;
+          if (wallDoc.door === CONST.WALL_DOOR_TYPES.SECRET) {
+
+            // If the door doesn't have a stealthy flag, only GMs can see like Foundry wants
+            const dc = wallDoc.flags.stealthy?.dc;
+            if (dc === undefined || dc === null) {
+              if (!game.user.isGM) return false;
+            }
+
+            // Otherwise, find the controlled tokens
+            else {
+              let tokens = canvas.tokens.controlled;
+              if (!tokens.length) {
+                if (!game.user.isGM) {
+                  tokens = canvas.scene.tokens.filter(t => {
+                    const userId = game.user.id;
+                    const ownership = t.actor.ownership[userId] ?? t.actor.ownership.default;
+                    return ownership >= 2;
+                  });
+                  if (!tokens.length) return false;
+                }
+              }
+
+              // Players only see secret doors if they control one unit
+              if (tokens.length === 1) {
+                const engine = game.stealthy.engine;
+                if (!engine.doorIsSpotted(this, tokens)) return false;
+                Stealthy.log(`Secret door DC${wallDoc.flags.stealthy?.dc} is found`, tokens);
+              }
+              else if (!game.user.isGM) return false;
+            }
+          }
+
+          // We've handled the GM check, do all the other stuff Foundry wants
+          return StealthyBaseEngine.BaseDoorControlIsVisibleSansGmCheck(this);
+        },
+        libWrapper.OVERRIDE,
+        { perf_mode: libWrapper.PERF_AUTO }
+      );
+
+      libWrapper.register(
+        Stealthy.MODULE_ID,
+        "WallConfig.prototype._updateObject",
+        async function (wrapped, event, formData) {
+          await wrapped(event, formData);
+          Stealthy.log(formData.dc);
+          const updateData =
+          {
+            flags:
+            {
+              stealthy:
+              {
+                dc: formData.dc,
+              }
+            }
+          };
+          let ids = this.editTargets;
+          if (ids.length == 0) {
+            ids = [this.object.id];
+          }
+
+          // Update all the edited walls
+          const updateDataset = ids.map(id => { return { _id: id, ...updateData }; });
+          const updateResult = await canvas.scene.updateEmbeddedDocuments("Wall", updateDataset);
+
+          return updateResult;
+        },
+        "WRAPPER"
+      );
+
+      // Inject custom settings into the wall config diallog
+      Hooks.on("renderWallConfig", (wallConfig, html, css) => {
+        Stealthy.log('renderWallConfig', { wallConfig, html, css });
+        if (css.document.door == 2) {
+          const dcBlock = `
+                  <div class="form-group">
+                    <label for="generateKey">${game.i18n.localize("stealthy.door.dc")}</label>
+                    <input type="number" name="dc"/ value="${css.object.flags.stealthy?.dc}">
+                  </div>`;
+          html.find(".form-group").last().after(dcBlock);
+
+          // Force config window to resize
+          wallConfig.setPosition({ height: "auto" });
+        }
+      });
+    }
   }
 
   testStealth(visionSource, target) {
@@ -218,12 +322,43 @@ export class StealthyBaseEngine {
     effect.flags.stealthy = flag;
     await actor.updateEmbeddedDocuments('ActiveEffect', [effect]);
   }
+
+  static BaseWallCreateDoorControlSansGmCheck(wall) {
+    wall.doorControl = canvas.controls.doors.addChild(new DoorControl(wall));
+    wall.doorControl.draw();
+    return wall.doorControl;
+  }
+
+  static BaseDoorControlIsVisibleSansGmCheck(doorControl) {
+    // Test two points which are perpendicular to the door midpoint
+    const w = doorControl.wall;
+    const ray = w.toRay();
+    const [x, y] = w.midpoint;
+    const [dx, dy] = [-ray.dy, ray.dx];
+    const t = 3 / (Math.abs(dx) + Math.abs(dy)); // Approximate with Manhattan distance for speed
+    const points = [
+      { x: x + (t * dx), y: y + (t * dy) },
+      { x: x - (t * dx), y: y - (t * dy) }
+    ];
+
+    // Test each point for visibility
+    return points.some(p => {
+      return canvas.effects.visibility.testVisibility(p, { object: doorControl, tolerance: 0 });
+    });
+  }
+
+  doorIsSpotted(doorControl, tokens) {
+    const dc = doorControl.wall.document.flags.stealthy.dc;
+    const spotter = tokens[0].actor;
+    const { value: perception } = this.getSpotFlagAndValue(spotter, this.findSpotEffect(spotter));
+    return perception >= dc;
+  }
 }
 
 export class Stealthy {
 
   static MODULE_ID = 'stealthy';
-  
+
   constructor(makeEngine) {
     this.engine = makeEngine();
     this.engine.patchFoundry();